BEO-11088 Media marks (#1016)
Motivation:
in the current solution, we use couchbase bucket "mediamarks" for reading more, than for writing: 
- read user-history `u{}`  and retrieve list of assetIds
- then for each assetId  read user-asset-history `u{}_m{}`(`u{}_epg{}`/`u{}_npvr{}`) which contains play location.  
so if we call `assethistory.list` and user have 100 assets in history, then we make 1+100 calls to couchbase.

Idea:
if we save to user-history `u{}` not assetIds only, but also the last play location(and other HIT props), then we don't need to read from user-asset-history `u{}_m{}`.

user-asset-history `u{}_m{}` was introduced to make a fast update for HIT requests. there is a lot of HIT requests during watching a movie, BUT we have CCProxy, which sends the last(!) HIT only. So the number of HIT requests to Phoenix is similar to the number of FIRST_PLAY/STOP/PAUSE requests. So it shouldn't be a performance issue to read full user history in order to update only one asset location in it.

Modifications: 
- save more info to `AssetAndLocation`
- update user history(Media,EPG,NPVR) not only for first-play
- try read play location from user-history, if not exist - from user-asset-history
- feature flag "mediamarks-play-location-in-user-object"

Future:
need to change DB where we store user history: key-value doesn't suit use cases